refactor: split handler.go and writes.go further to meet the ~800-line target
#237's acceptance criteria include "No file over ~800 lines in these
packages" — missed in the first pass, which left handler.go at 1511
lines (untouched) and writes.go at 1230 lines after the initial split.

handler.go (1511 -> 763):
- compat.go: extract ServeHTTP's inline interceptor blocks (interactive
  sub-resource rejection, k9s SelfSubjectAccessReview/RulesReview
  shims) into named methods. This is the one non-trivial move: the
  logic lived inline in ServeHTTP's body, not as separate top-level
  declarations, so it's an extract-method refactor rather than pure
  code motion — behavior is identical, only the call site changed from
  inline code to a function call.
- discovery.go: the /api, /apis, /apis/<group>[/<version>] discovery-
  serving functions (tryServeFromStore through serveGroupResourceList).
- status.go: writeStatus, writeJSON, statusObj, notFoundStatus,
  watchTimeout, bookmarkResourceVersion.
- scale.go (existing file): serveScale, alongside the write-side scale
  handling it's thematically paired with.
- table_row.go: extractTableRow (deliberately not added to table.go,
  already at 980 lines and outside #237's scope).

writes.go (1230 -> 650):
- scheduling.go: the node/pod scheduling shim (ensureSchedulableNode,
  schedulePod, knownNodeNames*, synthesizeNode*, ensurePodStatusPending,
  podNodeName, setSpecNodeName).
- crd.go: the CRD-creation synthesis (ensureCRDEstablished,
  registerCRDResourceInfo).
- write_util.go: generic path/metadata parsing helpers (patchMediaType,
  supportedPatchType, kindForResource, stampTypeMeta, allowedMethods,
  listPathFor, namespacesIsScope, parseWritePath, metaString, metaInt,
  dropDeletedNamespaceItems, isJSONObject, mergeMeta, protectSpecOnly,
  replaceField).

table.go (980 lines) is the one remaining file over ~800 in these
packages — pre-existing, never mentioned in #237's Problem section,
and out of scope.

No logic changes; every moved function/type kept its exact body and
doc comment.
