+++ internal/server/writes.go
@@ -5,15 +5,12 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/google/uuid"
-	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/yaml"
 
 	kstore "github.com/phenixblue/k8shark/internal/store"
@@ -232,10 +229,6 @@ func (h *handler) synthesizeOverlayObject(resource, namespace, name, base string
 	h.overlay.storeIfAbsent("", "v1", resource, namespace, name, obj, rv)
 }
 
-// defaultSyntheticNode is the node synthesized for scheduling when the capture
-// contains none.
-const defaultSyntheticNode = "kwok-node-0"
-
 // syncEpoch applies the overlay's reset-on-loop and, when a reset occurred,
 // re-synthesizes the scheduling node if needed. The synthetic node lives in the
 // overlay, so a loop wrap would otherwise drop it — leaving a nodeless capture
@@ -248,289 +241,6 @@ func (h *handler) syncEpoch() {
 	if h.overlay.syncEpoch(h.clock) && h.schedulePods {
 		h.ensureSchedulableNode()
 	}
-}
-
-// ensureSchedulableNode synthesizes a node when the capture has none, so a
-// scheduling target — and a node for KWOK to manage — exists from startup rather
-// than only appearing when the first Pod is created. Idempotent: synthesizeNode
-// is a no-op when the node already exists.
-//
-// Node presence is evaluated at the window END, not the current clock instant.
-// A capture's first /api/v1/nodes snapshot commonly lands a few seconds after
-// the window start (from ≈ metadata.CapturedAt, an approximation), so checking
-// at `from` would wrongly see "no nodes" at startup and synthesize kwok-node-0
-// for a capture that actually contains nodes (#172).
-func (h *handler) ensureSchedulableNode() {
-	at := h.at
-	if h.clock != nil {
-		_, at = h.clock.Window()
-	}
-	if len(h.knownNodeNamesAt(at)) == 0 {
-		h.synthesizeNode(defaultSyntheticNode)
-	}
-}
-
-// schedulePod binds an unscheduled Pod to a node — the scheduler replay lacks —
-// picking round-robin over the known nodes (captured + overlay) and synthesizing
-// a KWOK-managed node if none exist. Returns the body with spec.nodeName set.
-func (h *handler) schedulePod(body json.RawMessage) json.RawMessage {
-	nodes := h.knownNodeNames()
-	if len(nodes) == 0 {
-		h.synthesizeNode(defaultSyntheticNode)
-		nodes = []string{defaultSyntheticNode}
-	}
-	// Take the modulo in int64, then convert the bounded [0,len) result — casting
-	// the raw counter to int first could overflow negative on a 32-bit platform.
-	idx := h.overlay.nextScheduleIndex() % int64(len(nodes))
-	return setSpecNodeName(body, nodes[int(idx)])
-}
-
-// knownNodeNames returns the sorted names of Nodes visible in writable replay:
-// those reconstructed from the capture as-of the clock, merged with the overlay
-// (overlay-created nodes added, tombstoned ones removed).
-func (h *handler) knownNodeNames() []string {
-	at := h.at
-	if h.clock != nil {
-		at = h.clock.Now()
-	}
-	return h.knownNodeNamesAt(at)
-}
-
-// knownNodeNamesAt returns the sorted node names visible as-of the given instant
-// (captured nodes reconstructed at `at`, merged with overlay writes).
-func (h *handler) knownNodeNamesAt(at time.Time) []string {
-	var items []json.RawMessage
-	if body, code, err := h.store.ReconstructAt("/api/v1/nodes", at); err == nil && code == 200 {
-		var l struct {
-			Items []json.RawMessage `json:"items"`
-		}
-		if json.Unmarshal(body, &l) == nil {
-			items = l.Items
-		}
-	}
-	items, _ = h.overlay.applyToList("", "v1", "nodes", "", items)
-	var names []string
-	for _, it := range items {
-		if n := metaString(it, "name"); n != "" {
-			names = append(names, n)
-		}
-	}
-	sort.Strings(names)
-	return names
-}
-
-// synthesizeNode stores a synthetic Ready Node, annotated so a stock `kwok` run
-// manages it (`kwok.x-k8s.io/node: fake`). It gives the scheduling shim a target
-// when the capture has no nodes, and KWOK a node to keep Ready.
-func (h *handler) synthesizeNode(name string) {
-	h.synthesizeOverlayObject("nodes", "", name, syntheticNodeBase(name))
-}
-
-// syntheticNodeBase is the base body for a synthesized Node (metadata name/uid/rv
-// are stamped by synthesizeOverlayObject). Built via json.Marshal so the node
-// name is always correctly escaped.
-func syntheticNodeBase(name string) string {
-	obj := map[string]any{
-		"apiVersion": "v1",
-		"kind":       "Node",
-		"metadata": map[string]any{
-			"annotations": map[string]any{"kwok.x-k8s.io/node": "fake"},
-			"labels": map[string]any{
-				"type":                   "kwok",
-				"kubernetes.io/os":       "linux",
-				"kubernetes.io/hostname": name,
-			},
-		},
-		"spec": map[string]any{},
-		"status": map[string]any{
-			"phase":       "Running",
-			"conditions":  []any{map[string]any{"type": "Ready", "status": "True", "reason": "KubeletReady"}},
-			"allocatable": map[string]any{"cpu": "32", "memory": "256Gi", "pods": "110"},
-			"capacity":    map[string]any{"cpu": "32", "memory": "256Gi", "pods": "110"},
-		},
-	}
-	b, _ := json.Marshal(obj)
-	return string(b)
-}
-
-// ensurePodStatusPending sets status.phase=Pending on a pod body when it has no
-// phase, mirroring the apiserver's create-time default. Returns body unchanged on
-// a decode/encode error or if a phase is already set.
-func ensurePodStatusPending(body json.RawMessage) json.RawMessage {
-	var m map[string]any
-	if err := json.Unmarshal(body, &m); err != nil || m == nil {
-		return body
-	}
-	status, ok := m["status"].(map[string]any)
-	if !ok || status == nil {
-		status = map[string]any{}
-		m["status"] = status
-	}
-	if p, _ := status["phase"].(string); p != "" {
-		return body // already has a phase; leave it
-	}
-	status["phase"] = "Pending"
-	out, err := json.Marshal(m)
-	if err != nil {
-		return body
-	}
-	return out
-}
-
-// ensureCRDEstablished synthesizes the status a real apiextensions-apiserver
-// stamps on a newly created CustomResourceDefinition — NamesAccepted and
-// Established conditions, acceptedNames, and storedVersions — matching what
-// kstatus (and `kubectl wait --for condition=Established`) look for. now is
-// used as every timestamp field's value (this is a one-shot synthesis at
-// creation, not a real transition history).
-func ensureCRDEstablished(body json.RawMessage, now string) json.RawMessage {
-	var m map[string]any
-	if err := json.Unmarshal(body, &m); err != nil || m == nil {
-		return body
-	}
-	spec, _ := m["spec"].(map[string]any)
-	names, ok := spec["names"].(map[string]any)
-	if !ok || names == nil {
-		names = map[string]any{} // non-nil: a real CRD's acceptedNames is never omitted
-	}
-	storedVersions := []string{} // non-nil: a real CRD's storedVersions is never omitted
-	if versions, ok := spec["versions"].([]any); ok {
-		for _, v := range versions {
-			vm, ok := v.(map[string]any)
-			if !ok {
-				continue
-			}
-			if storage, _ := vm["storage"].(bool); storage {
-				if n, ok := vm["name"].(string); ok {
-					storedVersions = append(storedVersions, n)
-				}
-			}
-		}
-	} else if v, ok := spec["version"].(string); ok && v != "" {
-		// Legacy apiextensions.k8s.io/v1beta1 CRDs could specify a single
-		// top-level spec.version instead of the spec.versions list v1
-		// introduced; treat it as the (only) stored version.
-		storedVersions = append(storedVersions, v)
-	}
-	m["status"] = map[string]any{
-		"acceptedNames": names,
-		"conditions": []map[string]any{
-			{
-				"type": "NamesAccepted", "status": "True",
-				"reason": "NoConflicts", "message": "no conflicts found",
-				"lastTransitionTime": now,
-			},
-			{
-				"type": "Established", "status": "True",
-				"reason": "InitialNamesAccepted", "message": "the initial names have been accepted",
-				"lastTransitionTime": now,
-			},
-		},
-		"storedVersions": storedVersions,
-	}
-	out, err := json.Marshal(m)
-	if err != nil {
-		return body
-	}
-	return out
-}
-
-// registerCRDResourceInfo parses a freshly created CustomResourceDefinition's
-// spec and registers its defined group/version/resource/kind with the store's
-// discovery metadata (kstore.CaptureStore.MergeResourceInfo), so /apis and
-// /apis/<group>/<version> — and therefore `kubectl api-resources` and any
-// client that walks discovery (e.g. `istioctl analyze`) — reflect it
-// immediately, rather than only once the archive is reloaded. Best-effort: a
-// malformed/unparseable CRD body is silently skipped rather than failing the
-// create — the CRD object itself is still stored either way; only its
-// discovery visibility is affected.
-func (h *handler) registerCRDResourceInfo(body json.RawMessage) {
-	type crdVersion struct {
-		Name   string `json:"name"`
-		Served bool   `json:"served"`
-	}
-	var crd struct {
-		Spec struct {
-			Group string `json:"group"`
-			Names struct {
-				Kind       string   `json:"kind"`
-				Singular   string   `json:"singular"`
-				Plural     string   `json:"plural"`
-				ShortNames []string `json:"shortNames"`
-			} `json:"names"`
-			Scope    string       `json:"scope"`
-			Versions []crdVersion `json:"versions"`
-			Version  string       `json:"version"` // legacy apiextensions.k8s.io/v1beta1
-		} `json:"spec"`
-	}
-	if err := json.Unmarshal(body, &crd); err != nil {
-		return
-	}
-	if crd.Spec.Group == "" || crd.Spec.Names.Plural == "" || crd.Spec.Names.Kind == "" {
-		// A missing Kind would otherwise fall through to MergeResourceInfo's
-		// kstore.ResourceToKind heuristic fallback, which is a plural-depluralizing
-		// guess known to be wrong for most CRDs — skipping keeps a malformed
-		// CRD body from polluting discovery with a made-up Kind.
-		return
-	}
-	// MergeResourceInfo treats namespaced as authoritative and overwrites any
-	// existing value, so an empty/unrecognized spec.scope (a malformed body)
-	// must be skipped entirely rather than defaulting to namespaced=true.
-	var namespaced bool
-	switch crd.Spec.Scope {
-	case "Namespaced":
-		namespaced = true
-	case "Cluster":
-		namespaced = false
-	default:
-		return
-	}
-	versions := crd.Spec.Versions
-	if len(versions) == 0 && crd.Spec.Version != "" {
-		// Legacy v1beta1 CRDs specify a single top-level spec.version instead
-		// of the spec.versions list v1 introduced; treat it as the (only)
-		// served version, mirroring ensureCRDEstablished's storedVersions
-		// fallback above.
-		versions = []crdVersion{{Name: crd.Spec.Version, Served: true}}
-	}
-	for _, v := range versions {
-		if !v.Served || v.Name == "" {
-			continue
-		}
-		h.store.MergeResourceInfo(crd.Spec.Group, v.Name, crd.Spec.Names.Plural, namespaced,
-			crd.Spec.Names.Kind, crd.Spec.Names.Singular, crd.Spec.Names.ShortNames)
-	}
-}
-
-// podNodeName returns a pod body's spec.nodeName ("" if unset).
-func podNodeName(body json.RawMessage) string {
-	var p struct {
-		Spec struct {
-			NodeName string `json:"nodeName"`
-		} `json:"spec"`
-	}
-	_ = json.Unmarshal(body, &p)
-	return p.Spec.NodeName
-}
-
-// setSpecNodeName returns body with spec.nodeName set to node, preserving the
-// rest of the object. On a decode/encode error the body is returned unchanged.
-func setSpecNodeName(body json.RawMessage, node string) json.RawMessage {
-	var m map[string]any
-	if err := json.Unmarshal(body, &m); err != nil || m == nil {
-		return body
-	}
-	spec, ok := m["spec"].(map[string]any)
-	if !ok || spec == nil {
-		spec = map[string]any{}
-		m["spec"] = spec
-	}
-	spec["nodeName"] = node
-	out, err := json.Marshal(m)
-	if err != nil {
-		return body
-	}
-	return out
 }
 
 func (h *handler) overlayReplace(w http.ResponseWriter, r *http.Request, group, version, resource, namespace, name, sub string) {
@@ -938,293 +648,3 @@ func (h *handler) nowRFC3339() string {
 
 // maxWriteBytes caps request bodies accepted by the overlay.
 const maxWriteBytes = 8 << 20 // 8 MiB
-
-// patchMediaType strips any parameters from a PATCH Content-Type and lower-cases
-// it (media types are case-insensitive per RFC 7231).
-func patchMediaType(contentType string) string {
-	ct := contentType
-	if i := strings.IndexByte(ct, ';'); i >= 0 {
-		ct = ct[:i]
-	}
-	return strings.ToLower(strings.TrimSpace(ct))
-}
-
-// supportedPatchType reports whether the PATCH Content-Type is one we handle;
-// an unknown or empty type is rejected with 415 rather than silently merged.
-func supportedPatchType(contentType string) bool {
-	switch patchMediaType(contentType) {
-	case "application/merge-patch+json", "application/json-patch+json",
-		"application/strategic-merge-patch+json", "application/apply-patch+yaml":
-		return true
-	}
-	return false
-}
-
-// kindForResource resolves a plural resource name to its registered built-in Kind
-// by inverting the apiserver's own kind→resource convention over the scheme's
-// known types. This is exact for every registered type (e.g. endpointslices →
-// EndpointSlice), unlike a name-capitalization heuristic. ok is false when no
-// built-in type in the group/version maps to the resource (custom resources).
-func kindForResource(gv schema.GroupVersion, resource string) (schema.GroupVersionKind, bool) {
-	for kind := range scheme.Scheme.KnownTypes(gv) {
-		gvk := gv.WithKind(kind)
-		if plural, _ := meta.UnsafeGuessKindToResource(gvk); plural.Resource == resource {
-			return gvk, true
-		}
-	}
-	return schema.GroupVersionKind{}, false
-}
-
-// stampTypeMeta returns obj with apiVersion/kind stamped, for a resource that
-// maps to a known built-in Kind (a no-op via withKind if already present or
-// the Kind is unknown/custom). Reads already do this (see handler.go's
-// trySingleItemGet and watch_replay.go's withKind for streamed events); write
-// responses need it too — client-go's typed Update/UpdateStatus calls
-// round-trip an object fetched via Get/List, whose TypeMeta the apiserver
-// strips on read (a well-known client-go quirk), so the request body often
-// carries no kind/apiVersion at all. A real apiserver's response is always
-// fully typed regardless; without stamping it here, a typed client-go decoder
-// fails the response with `Object 'Kind' is missing` — exactly what broke the
-// deployment controller's DeploymentStatus update path, found via the
-// upstream conformance suite's "Deployment should run the lifecycle of a
-// Deployment" spec.
-func stampTypeMeta(group, version, resource string, obj json.RawMessage) json.RawMessage {
-	gvk, ok := kindForResource(schema.GroupVersion{Group: group, Version: version}, resource)
-	if !ok {
-		return obj
-	}
-	apiVersion := version
-	if group != "" {
-		apiVersion = group + "/" + version
-	}
-	return withKind(obj, apiVersion, gvk.Kind)
-}
-
-// allowedMethods returns the Allow-header value for a write path shape, used on
-// 405 responses (RFC 7231 §6.5.5): collection paths allow create; item paths
-// allow the full CRUD set; the status subresource is read + update (no delete);
-// any other subresource is read-only.
-func allowedMethods(name, sub string) string {
-	switch {
-	case name == "":
-		return "GET, HEAD, POST, DELETE"
-	case sub == "":
-		return "GET, HEAD, PUT, PATCH, DELETE"
-	case sub == "status", sub == "scale":
-		return "GET, HEAD, PUT, PATCH"
-	default:
-		return "GET, HEAD"
-	}
-}
-
-// listPathFor builds the canonical list path for a GVR + namespace.
-func listPathFor(group, version, resource, namespace string) string {
-	base := "/api/" + version
-	if group != "" {
-		base = "/apis/" + group + "/" + version
-	}
-	if namespace != "" {
-		return base + "/namespaces/" + namespace + "/" + resource
-	}
-	return base + "/" + resource
-}
-
-// namespacesIsScope reports whether a leading "namespaces" segment is the
-// namespace-scoping keyword (/namespaces/<ns>/<resource>/…) rather than the core
-// cluster-scoped "namespaces" resource itself (/api/v1/namespaces/<name>). rest
-// is guaranteed to start with "namespaces" and have >= 2 elements.
-func namespacesIsScope(group string, rest []string) bool {
-	if group != "" {
-		// Non-core groups have no core "namespaces" resource. Treat a leading
-		// "namespaces" as the scoping keyword only when a namespaced resource
-		// follows (.../namespaces/<ns>/<resource>); a bare .../namespaces/<name>
-		// is left as an item of a (hypothetical) grouped "namespaces" resource.
-		return len(rest) >= 3
-	}
-	switch len(rest) {
-	case 2: // /api/v1/namespaces/<name> → the namespace object
-		return false
-	case 3: // /api/v1/namespaces/<name>/{status,finalize} → object subresource;
-		//        /api/v1/namespaces/<ns>/<resource>       → namespaced list
-		return rest[2] != "status" && rest[2] != "finalize"
-	default: // 4+: /api/v1/namespaces/<ns>/<resource>/<name>[/<sub>]
-		return true
-	}
-}
-
-// parseWritePath parses a write target into GVR + namespace + name + subresource.
-// name is empty for list-level (create) paths.
-func parseWritePath(path string) (group, version, resource, namespace, name, subresource string) {
-	parts := strings.Split(strings.TrimPrefix(path, "/"), "/")
-	var rest []string
-	switch {
-	case len(parts) >= 3 && parts[0] == "api":
-		version = parts[1]
-		rest = parts[2:]
-	case len(parts) >= 4 && parts[0] == "apis":
-		group = parts[1]
-		version = parts[2]
-		rest = parts[3:]
-	default:
-		return
-	}
-	// rest is one of:
-	//   [resource] | [resource name] | [resource name sub]
-	//   [namespaces ns resource] | [... name] | [... name sub]
-	//
-	// A leading "namespaces" is the namespace-scoping keyword only when a real
-	// namespaced resource follows. In the core group "namespaces" is ALSO a
-	// cluster-scoped resource, so /api/v1/namespaces/<name>[/status|/finalize]
-	// targets a namespace object — not a namespaced path (see namespacesIsScope).
-	if len(rest) >= 2 && rest[0] == "namespaces" && namespacesIsScope(group, rest) {
-		namespace = rest[1]
-		rest = rest[2:]
-	}
-	switch len(rest) {
-	case 1:
-		resource = rest[0]
-	case 2:
-		resource, name = rest[0], rest[1]
-	case 3:
-		resource, name, subresource = rest[0], rest[1], rest[2]
-	}
-	return
-}
-
-// ── small JSON helpers ──────────────────────────────────────────────────────
-
-// metaString reads metadata.<field> as a string ("" if absent/non-string).
-func metaString(obj json.RawMessage, field string) string {
-	var m struct {
-		Metadata map[string]json.RawMessage `json:"metadata"`
-	}
-	if err := json.Unmarshal(obj, &m); err != nil {
-		return ""
-	}
-	raw, ok := m.Metadata[field]
-	if !ok {
-		return ""
-	}
-	var s string
-	if err := json.Unmarshal(raw, &s); err != nil {
-		return ""
-	}
-	return s
-}
-
-// metaInt reads metadata.<field> as an int64 (0 if absent/non-number).
-func metaInt(obj json.RawMessage, field string) int64 {
-	var m struct {
-		Metadata map[string]json.RawMessage `json:"metadata"`
-	}
-	if err := json.Unmarshal(obj, &m); err != nil {
-		return 0
-	}
-	raw, ok := m.Metadata[field]
-	if !ok {
-		return 0
-	}
-	var n int64
-	if err := json.Unmarshal(raw, &n); err != nil {
-		return 0
-	}
-	return n
-}
-
-// dropDeletedNamespaceItems removes items whose metadata.namespace is in dns
-// (see overlay.deletedNamespaces) — used to cascade a namespace delete into
-// read results and into deletecollection's item set, for both captured and
-// overlay-created items.
-func dropDeletedNamespaceItems(items []json.RawMessage, dns map[string]struct{}) []json.RawMessage {
-	if len(dns) == 0 {
-		return items
-	}
-	kept := items[:0]
-	for _, it := range items {
-		if _, gone := dns[metaString(it, "namespace")]; !gone {
-			kept = append(kept, it)
-		}
-	}
-	return kept
-}
-
-// isJSONObject reports whether b is a JSON object ("{...}"), rejecting null,
-// arrays, and scalars — so client write bodies can't be e.g. "null".
-func isJSONObject(b []byte) bool {
-	var m map[string]json.RawMessage
-	return json.Unmarshal(b, &m) == nil && m != nil
-}
-
-// mergeMeta returns obj with the given metadata fields set/overwritten. It is
-// nil-safe: a null object or null metadata is treated as an empty object.
-func mergeMeta(obj json.RawMessage, updates map[string]any) json.RawMessage {
-	var m map[string]json.RawMessage
-	if err := json.Unmarshal(obj, &m); err != nil {
-		return obj
-	}
-	if m == nil {
-		m = map[string]json.RawMessage{}
-	}
-	meta := map[string]json.RawMessage{}
-	if raw, ok := m["metadata"]; ok {
-		if err := json.Unmarshal(raw, &meta); err != nil || meta == nil {
-			meta = map[string]json.RawMessage{}
-		}
-	}
-	for k, v := range updates {
-		if s, ok := v.(string); ok && s == "" {
-			continue // don't write empty namespace/uid/etc.
-		}
-		b, err := json.Marshal(v)
-		if err != nil {
-			continue
-		}
-		meta[k] = b
-	}
-	m["metadata"], _ = json.Marshal(meta)
-	out, err := json.Marshal(m)
-	if err != nil {
-		return obj
-	}
-	return out
-}
-
-// protectSpecOnly returns next with .spec forced back to current's .spec,
-// otherwise unchanged — the status subresource's real apiserver semantics
-// (see e.g. pkg/registry/apps/deployment/strategy.go's
-// deploymentStatusStrategy.PrepareForUpdate): only .spec is universally
-// protected against a status-subresource write, while .status and .metadata
-// (annotations, and often labels — this varies slightly per resource type
-// upstream, but protecting only .spec is the one rule that holds everywhere)
-// pass through from the submitted body. This matters in practice: the
-// deployment controller sets `deployment.kubernetes.io/revision` on the
-// Deployment itself via UpdateStatus (a full-object PUT to .../status), not a
-// spec/metadata write — an earlier version of this code protected all of
-// metadata too, which silently dropped that annotation and broke revision
-// tracking for every Deployment reconciled by --with-controller-manager.
-func protectSpecOnly(next, current json.RawMessage) json.RawMessage {
-	return replaceField(next, "spec", current)
-}
-
-// replaceField returns base with top-level field set to the same field taken
-// from src.
-func replaceField(base json.RawMessage, field string, src json.RawMessage) json.RawMessage {
-	var b map[string]json.RawMessage
-	if err := json.Unmarshal(base, &b); err != nil || b == nil {
-		return base
-	}
-	var s map[string]json.RawMessage
-	if err := json.Unmarshal(src, &s); err != nil || s == nil {
-		return base
-	}
-	if v, ok := s[field]; ok {
-		b[field] = v
-	} else {
-		delete(b, field)
-	}
-	out, err := json.Marshal(b)
-	if err != nil {
-		return base
-	}
-	return out
-}
